+++ x-pack/plugins/security_solution/public/sourcerer/containers/hooks.test.tsx
@@ -10,7 +10,7 @@ import { act, renderHook } from '@testing-library/react-hooks';
 import { waitFor } from '@testing-library/react';
 import { Provider } from 'react-redux';
 
-import { getScopeFromPath, useInitSourcerer, useSourcererDataView } from '.';
+import { getScopeFromPath, useSourcererDataView } from '.';
 import { mockPatterns } from './mocks';
 import type { RouteSpyState } from '../../common/utils/route/types';
 import {
@@ -31,6 +31,7 @@ import * as source from '../../common/containers/source/use_data_view';
 import { sourcererActions } from '../store';
 import { useInitializeUrlParam, useUpdateUrlParam } from '../../common/utils/global_query_string';
 import { createSourcererDataView } from './create_sourcerer_data_view';
+import { useInitSourcerer } from './use_init_sourcerer';
 
 const mockRouteSpy: RouteSpyState = {
   pageName: SecurityPageName.overview,
